+++ components/UserJournals.jsx
@@ -1,13 +1,13 @@
 
 function UserJournals({ children, title, id, user, username, createdOn }) {
     return (
-        <div className="grid grid-cols-3 items-center font-semibold text-sm text-neutral-focus my-2
-            sm:text-base md:text-lg lg:text-xl hover:bg-accent/50 rounded-lg py-2 px-2">
-            <div className="truncate">{title}</div>
-            <div className="col-span-2 ">
-                <ul className="grid grid-cols-3 justify-items-center items-center truncate" key={id}>
-                    <li className="truncate">{username}</li>
-                    <li className="truncate">{new Date(createdOn.toMillis()).toDateString()}</li>
+        <div className="flex justify-between md:grid md:grid-cols-3 items-center font-semibold text-sm text-neutral-focus my-2 py-2 pl-2
+            sm:text-base md:text-lg lg:text-xl hover:bg-accent/50 rounded-lg ">
+            <div className="">{title}</div>
+            <div className="col-span-2 w-1/2 md:w-full">
+                <ul className="flex justify-center items-center md:grid md:grid-cols-3 md:justify-items-center gap-1" key={id}>
+                    <li className="hidden md:block">{username}</li>
+                    <li className="hidden md:block">{new Date(createdOn.toMillis()).toDateString().split(' ').slice(1).join(' ')}</li>
                     {children}
                 </ul>
             </div>
